requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
pymongo==4.6.0
mongoengine==0.27.0
python-decouple==3.8
requests==2.31.0
Pillow==10.1.0
django-extensions==3.2.3
bcrypt==4.1.2
